requirements.txt
Pillow>=10.0
pywin32>=306
opencv-python>=4.8
numpy>=1.24
PyYAML>=6.0
